Refactor App & Fix tests (#5)
Buffers still need work
