--- test/sample.pdf
@@ -1,198 +1,0 @@
-%PDF-1.3
-%‚„œ”
-
-1 0 obj
-<<
-/Type /Catalog
-/Outlines 2 0 R
-/Pages 3 0 R
->>
-endobj
-
-2 0 obj
-<<
-/Type /Outlines
-/Count 0
->>
-endobj
-
-3 0 obj
-<<
-/Type /Pages
-/Count 2
-/Kids [ 4 0 R 6 0 R ] 
->>
-endobj
-
-4 0 obj
-<<
-/Type /Page
-/Parent 3 0 R
-/Resources <<
-/Font <<
-/F1 9 0 R 
->>
-/ProcSet 8 0 R
->>
-/MediaBox [0 0 612.0000 792.0000]
-/Contents 5 0 R
->>
-endobj
-
-5 0 obj
-<< /Length 1074 >>
-stream
-2 J
-BT
-0 0 0 rg
-/F1 0027 Tf
-57.3750 722.2800 Td
-( A Simple PDF File ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 688.6080 Td
-( This is a small demonstration .pdf file - ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 664.7040 Td
-( just for use in the Virtual Mechanics tutorials. More text. And more ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 652.7520 Td
-( text. And more text. And more text. And more text. ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 628.8480 Td
-( And more text. And more text. And more text. And more text. And more ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 616.8960 Td
-( text. And more text. Boring, zzzzz. And more text. And more text. And ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 604.9440 Td
-( more text. And more text. And more text. And more text. And more text. ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 592.9920 Td
-( And more text. And more text. ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 569.0880 Td
-( And more text. And more text. And more text. And more text. And more ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 557.1360 Td
-( text. And more text. And more text. Even more. Continued on page 2 ...) Tj
-ET
-endstream
-endobj
-
-6 0 obj
-<<
-/Type /Page
-/Parent 3 0 R
-/Resources <<
-/Font <<
-/F1 9 0 R 
->>
-/ProcSet 8 0 R
->>
-/MediaBox [0 0 612.0000 792.0000]
-/Contents 7 0 R
->>
-endobj
-
-7 0 obj
-<< /Length 676 >>
-stream
-2 J
-BT
-0 0 0 rg
-/F1 0027 Tf
-57.3750 722.2800 Td
-( Simple PDF File 2 ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 688.6080 Td
-( ...continued from page 1. Yet more text. And more text. And more text. ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 676.6560 Td
-( And more text. And more text. And more text. And more text. And more ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 664.7040 Td
-( text. Oh, how boring typing this stuff. But not as boring as watching ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 652.7520 Td
-( paint dry. And more text. And more text. And more text. And more text. ) Tj
-ET
-BT
-/F1 0010 Tf
-69.2500 640.8000 Td
-( Boring.  More, a little more text. The end, and just as well. ) Tj
-ET
-endstream
-endobj
-
-8 0 obj
-[/PDF /Text]
-endobj
-
-9 0 obj
-<<
-/Type /Font
-/Subtype /Type1
-/Name /F1
-/BaseFont /Helvetica
-/Encoding /WinAnsiEncoding
->>
-endobj
-
-10 0 obj
-<<
-/Creator (Rave \(http://www.nevrona.com/rave\))
-/Producer (Nevrona Designs)
-/CreationDate (D:20060301072826)
->>
-endobj
-
-xref
-0 11
-0000000000 65535 f
-0000000019 00000 n
-0000000093 00000 n
-0000000147 00000 n
-0000000222 00000 n
-0000000390 00000 n
-0000001522 00000 n
-0000001690 00000 n
-0000002423 00000 n
-0000002456 00000 n
-0000002574 00000 n
-
-trailer
-<<
-/Size 11
-/Root 1 0 R
-/Info 10 0 R
->>
-
-startxref
-2714
-%%EOF
